Refactor: change file format JavaScript to TypeScript
- 파일 형식을 자바스크립트(js, jsx)에서 타입스크립트로(ts, tsx) 변경
