+++ src/components/Banner/BannerArticle.tsx
