+++ src/components/Banner/index.ts
