+++ src/components/SectionTabs/index.ts
