+++ src/utils/formattingText.ts
